Split auth0 guardian resource into multiple files
